+++ comicbox/metadata/filename.py
@@ -12,8 +12,7 @@ import re
 from logging import getLogger
 from pathlib import Path
 
-from parse import compile
-from parse import with_pattern
+from parse import compile, with_pattern
 
 from comicbox.metadata.comic_base import ComicBaseMetadata
 
@@ -23,12 +22,14 @@ LOG = getLogger(__name__)
 
 @with_pattern(r"#?(\d|½)+\.?\d*\w*")
 def issue(text):
+    """Issue number."""
     res = ComicBaseMetadata.parse_issue(text)
     return res
 
 
-@with_pattern(r"vo?l?\.? ?\d+")
+@with_pattern(r"v(?:ol)?\.? ?\d+")
 def volume(text):
+    """Volume number or year."""
     text = text.lstrip("vV")
     text = text.lstrip("oO")
     text = text.lstrip("lL")
@@ -37,16 +38,24 @@ def volume(text):
     return int(text)
 
 
-@with_pattern(r"\(\d\d\d\d\)")
+@with_pattern(r"\(\d{4}\)")
 def year(text):
+    """Year."""
     return int(text[1:-1])
 
 
-@with_pattern(r"\(?of \d+\)")
+@with_pattern(r"\(?of \d+\)?")
 def issue_count(text):
+    """Issue count suffix."""
     text = text.split()[1]
     text = text.rstrip(")")
     return int(text)
+
+
+@with_pattern(r"([^\.\s]*)$")
+def ext(text):
+    """Last File Extension."""
+    return text
 
 
 def compile_parsers(patterns):
@@ -65,6 +74,7 @@ def compile_parsers(patterns):
                     "volume": volume,
                     "year": year,
                     "issue_count": issue_count,
+                    "ext": ext,
                 },
             )
             for pattern in patterns
@@ -77,41 +87,45 @@ def compile_parsers(patterns):
 class FilenameMetadata(ComicBaseMetadata):
     """Extract metadata from the filename."""
 
-    ALL_FIELDS = set(["series", "volume", "issue", "issue_count", "year", "ext"])
+    ALL_FIELDS = frozenset(["series", "volume", "issue", "issue_count", "year", "ext"])
     FIELD_SCHEMA = {key: None for key in ALL_FIELDS}
     # The order of these patterns is very important as patterns farther down
     # match after patterns at the top.
 
     PATTERNS = (
-        "{series} {volume:volume} {issue:issue} {title} {year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {title} {year:year} {remainder}.{ext}",
-        "{series} {issue:issue} {year:year} {remainder}.{ext}",
+        "{series} {volume:volume} {issue:issue} {title} {year:year} {remainder}"
+        + ".{ext:ext}",
+        "{series} {volume:volume} {title} {year:year} {remainder}.{ext:ext}",
+        "{series} {issue:issue} {year:year} {remainder}.{ext:ext}",
         "{series} {issue:issue} {issue_count:issue_count} {year:year} "
-        "{remainder}.{ext}",
-        "{series} {volume:volume} {title} {year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {issue:issue} {title} {year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {year:year} {issue:issue} {title} {remainder}.{ext}",
-        "{series} {volume:volume}{garbage}{year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {title} {year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {year:year} {issue:issue} {remainder}.{ext}",
-        "{series} {volume:volume} {issue:issue} {title} {year:year} {remainder}.{ext}"
-        "{series} {volume:volume} {issue:issue}.{ext}"
-        "{series} {volume:volume} {title} {year:year} {remainder}.{ext}"
+        "{remainder}.{ext:ext}",
+        "{series} {volume:volume} {title} {year:year} {remainder}.{ext:ext}",
+        "{series} {volume:volume} {issue:issue} {title} {year:year} {remainder}"
+        + ".{ext:ext}",
+        "{series} {volume:volume} {year:year} {issue:issue} {title} {remainder}"
+        + ".{ext:ext}",
+        "{series} {volume:volume}{garbage}{year:year} {remainder}.{ext:ext}",
+        "{series} {volume:volume} {title} {year:year} {remainder}.{ext:ext}",
+        "{series} {volume:volume} {year:year} {remainder}.{ext:ext}",
+        "{series} {volume:volume} {year:year} {issue:issue} {remainder}.{ext:ext}",
+        "{series} {volume:volume} {issue:issue} {title} {year:year} {remainder}"
+        + ".{ext:ext}"
+        "{series} {volume:volume} {issue:issue}.{ext:ext}"
+        "{series} {volume:volume} {title} {year:year} {remainder}.{ext:ext}"
         "{series} {issue:issue} {issue_count:issue_count} {year:year} "
-        "{remainder}.{ext}",
-        "{series} {issue:issue} {issue_count:issue_count} {remainder}.{ext}",
-        "{series} {issue:issue} {year:year}.{ext}",
-        "{series} {issue:issue} {year:year} {remainder}.{ext}",
-        "{series} {year:year} {issue:issue} {remainder}.{ext}",
-        "{series} {year:year} {remainder}.{ext}",
-        "{series} {volume:volume} {issue:issue}.{ext}",
-        "{series} {volume:volume} {issue:issue} {remainder}.{ext}",
-        "{series} {issue:issue} {remainder}.{ext}",
-        "{series} {issue:issue}.{ext}",
-        "{series}.{ext}",
-        "{issue:issue} {series}.{ext}",
-        "{issue:issue} {series} {remainder}.{ext}",
+        "{remainder}.{ext:ext}",
+        "{series} {issue:issue} {issue_count:issue_count} {remainder}.{ext:ext}",
+        "{series} {issue:issue} {year:year}.{ext:ext}",
+        "{series} {issue:issue} {year:year} {remainder}.{ext:ext}",
+        "{series} {year:year} {issue:issue} {remainder}.{ext:ext}",
+        "{series} {year:year} {remainder}.{ext:ext}",
+        "{series} {volume:volume} {issue:issue}.{ext:ext}",
+        "{series} {volume:volume} {issue:issue} {remainder}.{ext:ext}",
+        "{series} {issue:issue} {remainder}.{ext:ext}",
+        "{series} {issue:issue}.{ext:ext}",
+        "{series}.{ext:ext}",
+        "{issue:issue} {series}.{ext:ext}",
+        "{issue:issue} {series} {remainder}.{ext:ext}",
     )
 
     PATTERN_MAX_MATCHES = tuple([pattern.count("}") for pattern in PATTERNS])
@@ -134,10 +148,6 @@ class FilenameMetadata(ComicBaseMetadata):
         """Try one parser and return the results as a dict."""
         res = parser.parse(fn)
         if res:
-            # For testing new patterns
-            # from pprint import pprint
-
-            # pprint(res)
             return res.named
         return {}