+++ comicbox/run.py
@@ -2,11 +2,15 @@
 import os
 import sys
 
+from logging import getLogger
 from pathlib import Path
 from pprint import pprint
 
 from comicbox.comic_archive import ComicArchive
 from comicbox.version import VERSION
+
+
+LOG = getLogger(__name__)
 
 
 class Runner:
@@ -17,6 +21,7 @@ class Runner:
     def __init__(self, config):
         """Initialize actions and config."""
         self.config = config
+        self.noop = True
 
     def run_on_file(self, path):
         """Run operations on one file."""
@@ -30,20 +35,28 @@ class Runner:
         car = ComicArchive(path, config=self.config)
         if self.config.raw:
             car.print_raw()
+            self.noop = False
         if self.config.print:
             pprint(car.get_metadata())
+            self.noop = False
         if self.config.covers:
             car.extract_cover_as(self.config.dest_path)
+            self.noop = False
         if self.config.index_from:
             car.extract_pages(self.config.index_from, self.config.dest_path)
+            self.noop = False
         if self.config.export:
             car.export_files()
+            self.noop = False
         if self.config.cbz or self.config.delete_tags:
             car.recompress()
+            self.noop = False
         if self.config.import_fn:
             car.import_file(self.config.import_fn)
+            self.noop = False
         if self.config.rename:
             car.rename_file()
+            self.noop = False
 
     def recurse(self, path):
         """Perform operations recursively on files."""
@@ -57,12 +70,16 @@ class Runner:
         for root, dirnames, filenames in os.walk(path):
             root_path = Path(root)
             for dirname in dirnames:
-                self.recurse(dirname)
+                full_path = root_path / dirname
+                self.recurse(full_path)
             for filename in sorted(filenames):
                 if Path(filename).suffix.lower() not in self.SUFFIXES:
                     continue
                 full_path = root_path / filename
-                self.run_on_file(full_path)
+                try:
+                    self.run_on_file(full_path)
+                except Exception as ex:
+                    LOG.error(f"{full_path}: {ex}")
 
     def run(self):
         """Run actions with config."""
@@ -77,3 +94,6 @@ class Runner:
 
         for path in self.config.paths:
             self.run_on_file(Path(path))
+
+        if self.noop:
+            print("No action performed")