+++ fix-lint.sh
@@ -3,4 +3,5 @@ set -euxo pipefail
 poetry run isort --color .
 poetry run black .
 prettier --write .
-shfmt -s -w -i 4 ./*.sh ./**/*.sh
+# shfmt -s -w -i 4 ./*.sh ./**/*.sh
+shellharden --replace ./*.sh ./**/*.sh