+++ tests/test_comet.py
@@ -3,10 +3,7 @@ from decimal import Decimal
 
 from comicbox.metadata.comet import CoMet
 
-from .test_metadata import TEST_FILES_PATH
-from .test_metadata import TMP_ROOT
-from .test_metadata import read_metadata
-from .test_metadata import write_metadata
+from .test_metadata import TEST_FILES_PATH, TMP_ROOT, read_metadata, write_metadata
 
 
 ARCHIVE_PATH = TEST_FILES_PATH / "Captain Science #001-comet.cbz"
@@ -21,13 +18,13 @@ METADATA = {
     "day": 1,
     "volume": 1,
     "language": "en",
-    "characters": set(["Captain Science", "Gordon Dane"]),
+    "characters": frozenset(["Captain Science", "Gordon Dane"]),
     "credits": [
-        {"person": "Joe Orlando", "role": "writer"},
         {"person": "Wally Wood", "role": "penciller"},
+        {"person": "Joe Orlando", "role": "writer"},
     ],
     "ext": "cbz",
-    "genres": set(["Science Fiction"]),
+    "genres": frozenset(["Science Fiction"]),
     "description": "A long example description",
     "is_version_of": "Captain Science",
     "price": Decimal(0.10).quantize(Decimal("0.01")),
@@ -39,14 +36,15 @@ METADATA = {
     "title": "The Beginning",
     "last_mark": 12,
     "cover_image": "CaptainScience#1_01.jpg",
-    # "remainder": "comet",
     "page_count": 36,
 }
 
 
 def test_read_comet():
+    """Read comet metadata."""
     read_metadata(ARCHIVE_PATH, METADATA)
 
 
 def test_write_comet():
+    """Write comet metadata."""
     write_metadata(TMP_PATH, NEW_TEST_CBZ_PATH, METADATA, CoMet)